Move menu slice to entities
