+++ apps/client/src/app/rootReducer.ts
@@ -1,9 +1,9 @@
 import { combineReducers } from '@reduxjs/toolkit'
 
+import { menuSlice } from '@/entities/menu'
 import { baseApi } from '@/shared/api'
-import menuSlice from "@/app/layouts/menuSlice.ts";
 
 export const rootReducer = combineReducers({
   [baseApi.reducerPath]: baseApi.reducer,
-  menuSlice: menuSlice.reducer
+  [menuSlice.name]: menuSlice.reducer
 })
