+++ apps/client/src/entities/menu/index.ts
@@ -1,0 +1,1 @@
+export { menuSlice } from './model/menuSlice'
